requirements.txt
aws-cdk-lib==2.248.0
constructs>=10.0.0
boto3>=1.34.0
aws-lambda-powertools[all]>=2.40.0
